+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [0.22.0] - 2024-10-09
 
 ### Added
 
@@ -210,7 +210,8 @@
 
 - Initial release
 
-[0.21.1]: https://github.com/tinted-theming/tinty/compare/v0.20.0...v0.21.1
+[0.22.0]: https://github.com/tinted-theming/tinty/compare/v0.21.1...v0.22.0
+[0.21.1]: https://github.com/tinted-theming/tinty/compare/v0.21.0...v0.21.1
 [0.21.0]: https://github.com/tinted-theming/tinty/compare/v0.20.1...v0.21.0
 [0.20.1]: https://github.com/tinted-theming/tinty/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/tinted-theming/tinty/compare/v0.19.0...v0.20.0